Created collision system to take care of the collision logic
