+++ game-engine/src/App.tsx
@@ -1,13 +1,16 @@
 import "./App.css";
 import Game from "./games/flappyBird/Game";
 import { GameLoop } from "./engine/loop";
+import { CollisionSystem } from "./engine/Collision/CollisionSystem";
 
 function App() {
     return (
         <div id="game-window-div">
-            <GameLoop>
-                <Game />
-            </GameLoop>
+            <CollisionSystem>
+                <GameLoop>
+                    <Game />
+                </GameLoop>
+            </CollisionSystem>
         </div>
     );
 }
