+++ game-engine/src/engine/Collision/CollisionOptions.ts
@@ -1,0 +1,6 @@
+import { TypeGameObject } from "../types/objects/TypeGameObject";
+
+export type CollisionOptions = {
+    registerCollisionObjects(gameObject: TypeGameObject): void;
+    checkCollision(): void;
+};
