+++ game-engine/src/engine/Collision/useCollision.tsx
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { CollisionContext } from "./CollisionContext";
+
+export function useCollision() {
+    return useContext(CollisionContext);
+}
