--- game-engine/src/engine/CollisionSystem.ts
@@ -1,4 +1,0 @@
-export class CollisionSystem {
-    
-    
-}
