+++ game-engine/src/engine/loop/GameLoopContext.tsx
@@ -5,5 +5,5 @@ import { GameLoopOptions } from "./GameLoopOptions";
 export const GameLoopContext = createContext<GameLoopOptions>({
     registerObject(_gameObject: TypeGameObject) {},
     start() {},
-    registerCollisionObjects(_gameObject: TypeGameObject){}
+
 });
