+++ game-engine/src/engine/loop/GameLoopOptions.ts
@@ -3,5 +3,5 @@ import { TypeGameObject } from "../types/objects/TypeGameObject";
 export type GameLoopOptions = {
     registerObject(gameObject: TypeGameObject): void;
     start(): void;
-    registerCollisionObjects(gameObject: TypeGameObject): void;
+
 };
