+++ game-engine/src/games/flappyBird/Game.tsx
@@ -12,6 +12,7 @@ import { InputSystem } from "../../engine/input/InputSystem";
 import inputs from "../flappyBird/user_input.json"
 import { GameComponent } from "../../engine/superClasses/GameComponent";
 import React from "react";
+import { CollisionSystem } from "../../engine/Collision/CollisionSystem";
 
 
 export function pipeFactory() {
